Moved error handler into separate module.
