+++ package.json
@@ -18,7 +18,7 @@
   },
   "homepage": "https://github.com/irarainey/vscode-applicationregistrations/blob/main/README.md",
   "license": "SEE LICENSE IN LICENSE.md",
-  "version": "0.3.0",
+  "version": "0.4.0",
   "engines": {
     "vscode": "^1.73.0"
   },
