+++ src/services/app-role.ts
@@ -41,7 +41,7 @@ export class AppRoleService extends ServiceBase {
                 this.triggerOnComplete({ success: true, statusBarHandle: status });
             })
             .catch((error) => {
-                this.triggerOnError({ success: false, statusBarHandle: status, error: error, treeViewItem: item, previousIcon: previousIcon });
+                this.triggerOnError({ success: false, statusBarHandle: status, error: error, treeViewItem: item, previousIcon: previousIcon, treeDataProvider: this.treeDataProvider });
             });
     }
 
@@ -69,7 +69,7 @@ export class AppRoleService extends ServiceBase {
                 this.triggerOnComplete({ success: true, statusBarHandle: status });
             })
             .catch((error) => {
-                this.triggerOnError({ success: false, statusBarHandle: status, error: error, treeViewItem: item, previousIcon: previousIcon });
+                this.triggerOnError({ success: false, statusBarHandle: status, error: error, treeViewItem: item, previousIcon: previousIcon, treeDataProvider: this.treeDataProvider });
             });
     }
 
@@ -92,7 +92,7 @@ export class AppRoleService extends ServiceBase {
                 this.triggerOnComplete({ success: true, statusBarHandle: status });
             })
             .catch((error) => {
-                this.triggerOnError({ success: false, statusBarHandle: status, error: error, treeViewItem: item, previousIcon: previousIcon });
+                this.triggerOnError({ success: false, statusBarHandle: status, error: error, treeViewItem: item, previousIcon: previousIcon, treeDataProvider: this.treeDataProvider });
             });
     }
 
@@ -125,7 +125,7 @@ export class AppRoleService extends ServiceBase {
                     this.triggerOnComplete({ success: true, statusBarHandle: status });
                 })
                 .catch((error) => {
-                    this.triggerOnError({ success: false, statusBarHandle: status, error: error, treeViewItem: item, previousIcon: previousIcon });
+                    this.triggerOnError({ success: false, statusBarHandle: status, error: error, treeViewItem: item, previousIcon: previousIcon, treeDataProvider: this.treeDataProvider });
                 });
         }
     }
