+++ src/services/oauth2-permission-scope.ts
@@ -48,7 +48,7 @@ export class OAuth2PermissionScopeService extends ServiceBase {
                 this.triggerOnComplete({ success: true, statusBarHandle: status });
             })
             .catch((error) => {
-                this.triggerOnError({ success: false, statusBarHandle: status, error: error, treeViewItem: item, previousIcon: previousIcon });
+                this.triggerOnError({ success: false, statusBarHandle: status, error: error, treeViewItem: item, previousIcon: previousIcon, treeDataProvider: this.treeDataProvider });
             });
     }
 
@@ -76,7 +76,7 @@ export class OAuth2PermissionScopeService extends ServiceBase {
                 this.triggerOnComplete({ success: true, statusBarHandle: status });
             })
             .catch((error) => {
-                this.triggerOnError({ success: false, statusBarHandle: status, error: error, treeViewItem: item, previousIcon: previousIcon });
+                this.triggerOnError({ success: false, statusBarHandle: status, error: error, treeViewItem: item, previousIcon: previousIcon, treeDataProvider: this.treeDataProvider });
             });
     }
 
@@ -99,7 +99,7 @@ export class OAuth2PermissionScopeService extends ServiceBase {
                 this.triggerOnComplete({ success: true, statusBarHandle: status });
             })
             .catch((error) => {
-                this.triggerOnError({ success: false, statusBarHandle: status, error: error, treeViewItem: item, previousIcon: previousIcon });
+                this.triggerOnError({ success: false, statusBarHandle: status, error: error, treeViewItem: item, previousIcon: previousIcon, treeDataProvider: this.treeDataProvider });
             });
     }
 
@@ -132,7 +132,7 @@ export class OAuth2PermissionScopeService extends ServiceBase {
                     this.triggerOnComplete({ success: true, statusBarHandle: status });
                 })
                 .catch((error) => {
-                    this.triggerOnError({ success: false, statusBarHandle: status, error: error, treeViewItem: item, previousIcon: previousIcon });
+                    this.triggerOnError({ success: false, statusBarHandle: status, error: error, treeViewItem: item, previousIcon: previousIcon, treeDataProvider: this.treeDataProvider });
                 });
         }
     }
