+++ src/services/password-credential.ts
@@ -50,7 +50,7 @@ export class PasswordCredentialService extends ServiceBase {
                         window.showInformationMessage("New password copied to clipboard.", "OK");
                     })
                     .catch((error) => {
-                        this.triggerOnError({ success: false, statusBarHandle: status, error: error, treeViewItem: item, previousIcon: previousIcon });
+                        this.triggerOnError({ success: false, statusBarHandle: status, error: error, treeViewItem: item, previousIcon: previousIcon, treeDataProvider: this.treeDataProvider });
                     });
             }
         }
@@ -72,7 +72,7 @@ export class PasswordCredentialService extends ServiceBase {
                     this.triggerOnComplete({ success: true, statusBarHandle: status });
                 })
                 .catch((error) => {
-                    this.triggerOnError({ success: false, statusBarHandle: status, error: error, treeViewItem: item, previousIcon: previousIcon });
+                    this.triggerOnError({ success: false, statusBarHandle: status, error: error, treeViewItem: item, previousIcon: previousIcon, treeDataProvider: this.treeDataProvider });
                 });
         }
     }
