+++ src/types/app-reg-item-params.ts
@@ -1,8 +1,8 @@
 import { ThemeIcon, Command } from "vscode";
 import { AppRegItem } from "../models/app-reg-item";
 
-// This is the interface for the application registration tree view item parameters
-export interface AppRegItemParams {
+// This is the type for the application registration tree view item parameters
+export type AppRegItemParams = {
     label: string;
     context: string;
     value?: string;
@@ -18,4 +18,4 @@ export interface AppRegItemParams {
     tooltip?: string;
     order?: number;
     state?: boolean;
-}
+};
